requirements.txt
# dependencies
beautifulsoup4
colorama
md2pdf
python-dotenv
pyyaml
uvicorn
pydantic
fastapi
python-multipart
markdown
langchain>=0.2,<0.3
langchain_community>=0.2,<0.3
langchain-openai>=0.1,<0.2
boto3
langgraph
tiktoken
gpt-researcher
arxiv
PyMuPDF
requests
jinja2
aiofiles
mistune
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair
json5
loguru

# uncomment for testing
# pytest
# pytest-asyncio
